Feat: adjust controller update and delete with redis

+++ app/Http/Controllers/UserController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\PendingAction;
 use App\Models\User;
 use App\Models\Role; // Add this line to import the Role class
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
@@ -48,6 +50,10 @@ class UserController extends Controller
      */
     public function update(Request $request)
     {
+        // find the user
+        $user = User::findOrFail($request->id);
+
+        // Validate the incoming request data
         $validatedData = $request->validate([
             'id' => 'required|integer|exists:users,id',
             'username' => 'sometimes|string|max:255',
@@ -55,6 +61,22 @@ class UserController extends Controller
             'password' => 'sometimes|string',
             'role_id' => 'sometimes|integer|exists:role,id',
         ]);
+
+        // If the user is not an admin, save the update request to the pending actions table
+        if (Auth::user()->role_id == 2) {
+            $updateData = [
+                'id' => $user->id,
+                'validated_data' => $validatedData,
+                'type' => 'UnitKerja',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('update', $user->id, $updateData);
+
+            return redirect()->back()
+                ->with('success', 'Update request has been sent for approval.');
+        }
 
         $user = User::whereId($request->id)->firstOrFail();
 
@@ -72,14 +94,32 @@ class UserController extends Controller
      */
     public function destroy(Request $request)
     {
+        // Find the user
+        $user = User::findOrFail($request->id);
+
         $validatedData = $request->validate([
             'id' => 'required|integer|exists:users,id',
         ]);
 
-        $user = User::whereId($validatedData['id'])->firstOrFail();
+        // If the user is not an admin, save the delete request to the pending actions table
+        if (Auth::user()->role_id == 2) {
+            $deleteData = [
+                'id' => $user->id,
+                'username' => $user->username,
+                'email' => $user->email,
+                'role_id' => $user->role_id,
+                'type' => 'User',
+                'requested_by' => Auth::user()->id,
+                'requested_at' => now(),
+            ];
+
+            PendingAction::savePendingAction('delete', $user->id, $deleteData);
+
+            return redirect()->back()->with('success', 'Delete request has been sent for approval.');
+        }
+
         $user->delete();
 
         return redirect()->back()->with('success', 'User deleted successfully');
     }
-
 }
